requirements.txt
pandas>=2.0
openai>=1.0
python-dotenv>=1.0
streamlit>=1.28
